final psc

+++ pages/suppliers/suppliers.js
@@ -1,20 +1,24 @@
+function fetchSup() {
+  fetch('http://gamma.apexcode.ro/api/suppliers')
+    .then(resp => resp.json())
+    .then(renderSup)
+}
+const supCollection = document.getElementById('sup-collection')
+function renderSup(sup) {
+  supCollection.innerHTML = ""
+  sup.forEach(function (sup) {
+    supCollection.innerHTML += `
+ <div class="card" data-id=${sup.Id}>
+      <h2>${sup.Name}</h2>
+      <p>${sup.CUI}</p>
+      <button class="delete">Delete</button>
+ </div>
+`
+  })
+}
+fetchSup();
 
-var delId =1;
-
-fetch('http://gamma.apexcode.ro/api/suppliers')
-.then((resp) => resp.json())
-.then((resp) => {
-    resp.forEach((supplier) => {
-        document.getElementById('content').innerHTML += 
-        `<div>
-          <h3 ${supplier.Id}>${supplier.Name}</h3>
-          <p>${supplier.CUI}</p>
-          <a href="#" id="delete${delId}">Delete</a>
-        </div>`
-        delId++;
-    });
-})
-
+// adaugare supplier
 
 document.getElementById("save").addEventListener('click', addSupplier);
 function addSupplier() {
@@ -22,35 +26,37 @@ function addSupplier() {
   let cui = document.getElementById('cui').value;
 
 
-    var supplier = {
-        Name: name,
-        CUI: cui
-    };
+  var supplier = {
+    Name: name,
+    CUI: cui
+  };
 
-    fetch('http://gamma.apexcode.ro/api/suppliers', {
-        method: 'POST',
-        mode: 'cors',
-        cache: 'no-cache',
-        headers: {
-            'Accept': 'application/json',
-            'Content-Type': 'application/json'
-        },
-        body: JSON.stringify(supplier)
-    })
-        .then(function (response) {
-            console.log(response.statusText);
-        });
+  fetch("http://gamma.apexcode.ro/api/suppliers", {
+    method: 'POST',
+    mode: 'cors',
+    cache: 'no-cache',
+    headers: {
+      'Accept': 'application/json',
+      'Content-Type': 'application/json'
+    },
+    body: JSON.stringify(supplier)
+  })
+    .then(response => response.json())
+    .then(fetchCustomer)
 }
 
-document.getElementById(`delete${delId}`).addEventListener("click", function deleteSupplier(){
-    fetch(`http://gamma.apexcode.ro/api/suppliers/${delId}`, {
-      method: "DELETE" 
+// delete supplier
+
+document.getElementById('sup-collection').addEventListener('click', function (event) {
+  let delButton = event.target.className === "delete"
+  if (delButton) {
+    let id = event.target.parentElement.dataset.id
+    fetch(`http://gamma.apexcode.ro/api/suppliers/${id}`, {
+      method: 'DELETE'
     })
-      .then(function (response) {
-        return response.json();
-      })
-      .then(function () {
-        var supplierDeleted = document.querySelector('#name [value="' + id + '"]');
-        supplierDeleted.parentNode.removeChild(supplierDeleted);
-      });
-  }); 
+      .then(response => response.json())
+      .then(fetchSup)
+  }
+})
+
+
